restorevideo/requirements.txt
opencv-python==4.5.1
mtcnn==0.1.0
numpy==1.21.2
Pillow==8.3.1
git+https://github.com/google/deepdream.git
